updating util seed func

+++ internal/model/field.go
@@ -5,7 +5,7 @@ import "github.com/arthurlee945/Docrilla/internal/model/enum/field"
 type Field struct {
 	ID        string     `json:"id" db:"id"`
 	UUID      string     `json:"uuid" db:"uuid"`
-	ProjectID int        `json:"projectID" db:"project_id"`
+	ProjectID uint64     `json:"projectID" db:"project_id"`
 	X1        float64    `json:"x1" db:"x1"`
 	Y1        float64    `json:"y1" db:"y1"`
 	X2        float64    `json:"x2" db:"x2"`
